+++ Ejercicio43/Program.cs
@@ -12,11 +12,11 @@ namespace Ejercicio43
         public static void Main(string[] args)
         {
             int meses = 0;
-            int sumaDeVentas = 0;
-            int promedio = 0;
-            int sumaDeMesesQueSuperaronElPromedio = 0;
-            int contador = 0;
-            int PromedioDeMesesQueSuperaronElPromedioDeVentas = 0;
+            double sumaDeVentas = 0;
+            double promedio = 0;
+            double sumaDeMesesQueSuperaronElPromedio = 0;
+            double contador = 0;
+            double PromedioDeMesesQueSuperaronElPromedioDeVentas = 0;
 
 
             Console.WriteLine("ingrese los meses");
